+++ components/pages/home/hero.tsx
@@ -11,8 +11,8 @@ const Hero = (): React.ReactElement => {
     <div className="relative bg-gray-900">
       <div
         aria-hidden="true"
-        className="flex items-center justify-center h-4/6 mb-12 bg-fixed bg-center md:bg-cover"
-        style={{backgroundImage: "url('/images/background.jpg')"}}
+        className="flex items-center justify-center h-4/6 mb-12 bg-fixed bg-top md:bg-center md:bg-cover"
+        style={{backgroundImage: "url('/images/background-2.jpg')"}}
       >
         <div
           aria-hidden="true"
@@ -22,12 +22,13 @@ const Hero = (): React.ReactElement => {
         <div className="relative max-w-6xl mx-auto pt-32 pb-16 px-6 flex flex-col items-center text-center sm:pt-64 sm:pb-32 lg:px-0">
           <img
             src="/images/logo-white.svg"
-            alt=""
+            alt="Slave to Servant"
+            title="Slave to Servant"
             className="w-auto h-auto"
             style={{maxWidth: 300}}
           />
 
-          <h1 className="text-6xl font-king font-extrabold tracking-tight text-white lg:text-9xl mt-5 mb-10">
+          <h1 className="text-6xl font-berenika font-extrabold tracking-tight text-white lg:text-9xl mt-5 mb-10">
             Slave to Servant
           </h1>