+++ pages/_document.tsx
@@ -1,4 +1,4 @@
-import Document, {Html, Head, Main, NextScript} from 'next/document'
+import Document, {Head, Html, Main, NextScript} from 'next/document'
 import config from '../config'
 
 export default class PatrickFilmDocument extends Document {