+++ paypal_extends.d.ts
@@ -1,5 +1,3 @@
-import {PayPalButtonProps} from 'react-paypal-button-v2/dist/index'
-
 declare module 'react-paypal-button-v2/dist/index' {
   interface PayPalButtonProps {
     onCancel?: Function